requirements.txt
openai
requests
vobject
python-dotenv
pyyaml
